requirements.txt
asgiref==3.5.2
autopep8==1.7.0
bcrypt==4.0.0
boto3==1.24.63
botocore==1.27.63
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.1
coreapi==2.3.3
coreschema==0.0.4
cryptography==37.0.4
defusedxml==0.7.1
dj-rest-auth==2.2.4
Django==4.0
django-allauth==0.50.0
django-cors-headers==3.13.0
django-debug-toolbar==3.6.0
django-rest-auth==0.9.5
django-rest-authtoken==2.1.4
django-rest-knox==4.2.0
django-silk==5.0.1
django-storages==1.13.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
django-environ==0.9.0
drf-yasg==1.21.3
et-xmlfile==1.1.0
gprof2dot==2022.7.29
gunicorn==20.1.0
haversine==2.6.0
idna==3.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
jmespath==1.0.1
MarkupSafe==2.1.1
mysqlclient==2.1.1
oauthlib==3.2.0
openpyxl==3.0.10
packaging==21.3
Pillow==9.2.0
pycodestyle==2.9.1
pycparser==2.21
PyJWT==1.7.1
pyparsing==3.0.9
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2022.2.1
requests==2.28.1
requests-oauthlib==1.3.1
s3transfer==0.6.0
sentry-sdk==1.9.10
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
tzdata==2022.2
uritemplate==4.1.1
urllib3==1.26.12
django-redis==5.3.0
twilio==9.0.2